# Improve ansi.Parser performance and memory allocs