@aymanbagabas 2024-08-02
@ccoVeille I've added some benchmarks above. Looks like it takes slightly more time now, but less allocs in certain situations 🤔