@aymanbagabas 2024-08-02
Ok, a little better
```
goos: darwin
goarch: arm64
pkg: github.com/charmbracelet/x/ansi
                                      │   old.txt   │               new.txt               │
                                      │   sec/op    │   sec/op     vs base                │
Parser/simple-16                        80.91µ ± 1%   80.02µ ± 1%   -1.10% (p=0.035 n=10)
Parser/params-16                        83.31µ ± 0%   82.57µ ± 1%   -0.88% (p=0.035 n=10)
Parser/params_and_data-16               82.45µ ± 0%   83.67µ ± 0%   +1.48% (p=0.000 n=10)
ParserUTF8/simple-16                    77.71µ ± 0%   71.73µ ± 0%   -7.69% (p=0.000 n=10)
ParserUTF8/params-16                    76.61µ ± 0%   71.84µ ± 0%   -6.23% (p=0.000 n=10)
ParserUTF8/params_and_data-16           76.46µ ± 0%   71.68µ ± 0%   -6.26% (p=0.000 n=10)
ParserStateChanges/simple-16            207.5n ± 1%   198.3n ± 0%   -4.41% (p=0.000 n=10)
ParserStateChanges/params-16            222.5n ± 0%   197.8n ± 1%  -11.10% (p=0.000 n=10)
ParserStateChanges/params_and_data-16   207.3n ± 2%   198.5n ± 0%   -4.29% (p=0.000 n=10)
geomean                                 11.03µ        10.53µ        -4.57%

                                      │    old.txt     │                  new.txt                  │
                                      │      B/op      │     B/op      vs base                     │
Parser/simple-16                          0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
Parser/params-16                        4.133Ki ± 0%     0.000Ki ± 0%  -100.00% (p=0.000 n=10)
Parser/params_and_data-16                 0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserUTF8/simple-16                      0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserUTF8/params-16                      0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserUTF8/params_and_data-16             0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserStateChanges/simple-16              0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserStateChanges/params-16              8.000 ± 0%       0.000 ± 0%  -100.00% (p=0.000 n=10)
ParserStateChanges/params_and_data-16     0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
geomean                                              ²                 ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                                      │   old.txt    │                 new.txt                 │
                                      │  allocs/op   │ allocs/op   vs base                     │
Parser/simple-16                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
Parser/params-16                        121.0 ± 0%       0.0 ± 0%  -100.00% (p=0.000 n=10)
Parser/params_and_data-16               0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserUTF8/simple-16                    0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserUTF8/params-16                    0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserUTF8/params_and_data-16           0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserStateChanges/simple-16            0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ParserStateChanges/params-16            1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
ParserStateChanges/params_and_data-16   0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
geomean                                            ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean
```